requirements.txt
requests
pandas
python-dotenv
apache-airflow
sqlalchemy
psycopg2-binary